scripts/requirements.txt
# CLIP and ML dependencies
torch>=2.0.0
open-clip-torch>=2.20.0
transformers>=4.30.0

# Image processing
Pillow>=9.5.0

# Database
psycopg2-binary>=2.9.6

# Utilities
requests>=2.31.0
python-dotenv>=1.0.0
numpy>=1.24.0
